+++ docs/CHANGELOG.md
@@ -45,14 +45,12 @@ sections to include in release notes:
 - Split wood carbon pool into `plantWoodC` and `plantWoodCStorageDelta` to track NPP storage lag (#248)
 - New moisture dependency function option for soil respiration, controlled by the `anaerobic` cli option (#259)
 - New moisture dependency functions for N volatilization, methane (#259)
-<<<<<<< HEAD
 - Adds Nitrogen demand, fixation flux, and uptake flux (#265)
-=======
 - Methane production option, controlled by the `anaerobic` cli option (#269)
->>>>>>> master
 
 ### Fixed
 
+- Balance check now accounts for mass adjustments from non-negative stock clamping
 - Out of order error message for events (#166)
 
 ### Changed